Store 2D grids correctly.
This commit uses the sorting algorithm to sort 2D gpi and subsets
according to the lon, lat ordering that is preferred by netCDF. This is
especially important for the latitudes since they have to be stored from
90 to -90 in netCDF.

To accomodate this more flexible model of storing 2D grids it was
necessary to make a few additional changes. The gpidirect flag is no
longer taken into account since the reordering during storage most
likely changed the simple np.arange ordering that gpidirect assumes.

The tests for grid equality no longer look for exact array matches but
for same tuples of gpi, lon, lat and cell if it exitsts. The subsets are
now defined as being equal when they select the same gpis from the grid.