Remove clipping from lonlat2cell (#77)
When using the current implementation of `lonlat2cell` with cellsize < 1, the clipping leads to very large cells. For example, with `cellsize=0.5` all points with lat>0 and lon>0 will be mapped to the same cell, because in this case `(lat+90)/cellsize = (lat+90)*2 > 180` and similarly `(lon+180)/cellsize = (lon+180)*2 > 360`.
When using cellsize >= 1, clipping is only necessary if latitude and longitude are not in (-90, 90) and (-180, 180), respectively, but this is already enforced in the BasicGrid init, so it shouldn't be necessary here.